refactor: [M3-8765] - Remove unused Marketplace feature flags (#11133)
* remove unused flags

* rename `oneClickAppsv2.ts` to `oneClickApps.ts`

* add chageset and fix import

---------
