+++ packages/manager/.changeset/pr-11133-tech-stories-1729550913039.md
@@ -1,0 +1,5 @@
+---
+"@linode/manager": Tech Stories
+---
+
+Remove unused Marketplace feature flags ([#11133](https://github.com/linode/manager/pull/11133))
