+++ packages/manager/src/featureFlags.ts
@@ -1,4 +1,4 @@
-import type { Doc, OCA } from './features/OneClickApps/types';
+import type { OCA } from './features/OneClickApps/types';
 import type { TPAProvider } from '@linode/api-v4/lib/profile';
 import type { NoticeVariant } from 'src/components/Notice/Notice';
 
@@ -74,8 +74,6 @@ interface gpuV2 {
   planDivider: boolean;
 }
 
-type OneClickApp = Record<string, string>;
-
 interface DesignUpdatesBannerFlag extends BaseFeatureFlag {
   key: string;
   link: string;
@@ -115,8 +113,6 @@ export interface Flags {
   metadata: boolean;
   objMultiCluster: boolean;
   objectStorageGen2: BaseFeatureFlag;
-  oneClickApps: OneClickApp;
-  oneClickAppsDocsOverride: Record<string, Doc[]>;
   productInformationBanners: ProductInformationBannerFlag[];
   promos: boolean;
   promotionalOffers: PromotionalOffer[];
@@ -137,7 +133,7 @@ interface MarketplaceAppOverride {
   /**
    * Define app details that should be overwritten
    *
-   * If you are adding an app that is not already defined in "oneClickAppsv2.ts",
+   * If you are adding an app that is not already defined in "oneClickApps.ts",
    * you *must* include all required OCA properties or Cloud Manager could crash.
    *
    * Pass `null` to hide the marketplace app
