+++ packages/manager/src/features/Linodes/LinodeCreate/Tabs/Marketplace/utilities.ts
@@ -1,6 +1,6 @@
 import { decode } from 'he';
 
-import { oneClickApps } from 'src/features/OneClickApps/oneClickAppsv2';
+import { oneClickApps } from 'src/features/OneClickApps/oneClickApps';
 import { useFlags } from 'src/hooks/useFlags';
 import { useMarketplaceAppsQuery } from 'src/queries/stackscripts';
